requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12
aiofiles==24.1.0
numpy>=1.26.0
httpx==0.28.1
python-dotenv==1.1.1

